initial creation of HtApi serivce object
